requirements.txt
pyTelegramBotAPI==4.27.0
fastapi==0.108.0
uvicorn[standard]==0.23.1
python-dotenv==1.0.1
g4f>=0.3.0.7
pydantic>=2.0.0
requests>=2.31.0
aiosqlite

